+++ editor/src/marionette/variable-binding.ts
